commit for updae (ashad)
cakephp

+++ lib/Cake/Console/Templates/default/actions/controller_actions.ctp
@@ -36,11 +36,11 @@
  * @return void
  */
 	public function <?php echo $admin ?>view($id = null) {
-		if (!$this-><?php echo $currentModelName; ?>->exists($id)) {
+		$this-><?php echo $currentModelName; ?>->id = $id;
+		if (!$this-><?php echo $currentModelName; ?>->exists()) {
 			throw new NotFoundException(__('Invalid <?php echo strtolower($singularHumanName); ?>'));
 		}
-		$options = array('conditions' => array('<?php echo $currentModelName; ?>.' . $this-><?php echo $currentModelName; ?>->primaryKey => $id));
-		$this->set('<?php echo $singularName; ?>', $this-><?php echo $currentModelName; ?>->find('first', $options));
+		$this->set('<?php echo $singularName; ?>', $this-><?php echo $currentModelName; ?>->read(null, $id));
 	}
 
 <?php $compact = array(); ?>
@@ -91,7 +91,8 @@
  * @return void
  */
 	public function <?php echo $admin; ?>edit($id = null) {
-		if (!$this-><?php echo $currentModelName; ?>->exists($id)) {
+		$this-><?php echo $currentModelName; ?>->id = $id;
+		if (!$this-><?php echo $currentModelName; ?>->exists()) {
 			throw new NotFoundException(__('Invalid <?php echo strtolower($singularHumanName); ?>'));
 		}
 		if ($this->request->is('post') || $this->request->is('put')) {
@@ -108,8 +109,7 @@
 <?php endif; ?>
 			}
 		} else {
-			$options = array('conditions' => array('<?php echo $currentModelName; ?>.' . $this-><?php echo $currentModelName; ?>->primaryKey => $id));
-			$this->request->data = $this-><?php echo $currentModelName; ?>->find('first', $options);
+			$this->request->data = $this-><?php echo $currentModelName; ?>->read(null, $id);
 		}
 <?php
 		foreach (array('belongsTo', 'hasAndBelongsToMany') as $assoc):
@@ -131,17 +131,19 @@
 /**
  * <?php echo $admin ?>delete method
  *
- * @throws NotFoundException
  * @throws MethodNotAllowedException
+ * @throws NotFoundException
  * @param string $id
  * @return void
  */
 	public function <?php echo $admin; ?>delete($id = null) {
+		if (!$this->request->is('post')) {
+			throw new MethodNotAllowedException();
+		}
 		$this-><?php echo $currentModelName; ?>->id = $id;
 		if (!$this-><?php echo $currentModelName; ?>->exists()) {
 			throw new NotFoundException(__('Invalid <?php echo strtolower($singularHumanName); ?>'));
 		}
-		$this->request->onlyAllow('post', 'delete');
 		if ($this-><?php echo $currentModelName; ?>->delete()) {
 <?php if ($wannaUseSession): ?>
 			$this->Session->setFlash(__('<?php echo ucfirst(strtolower($singularHumanName)); ?> deleted'));
